Revert "rename files from bugXXXXXXX to something more descriptive"
Renaming the files messes up the version history.

This reverts commit a65b30987ba030c7698351b3afbc328b4eeb797b.
