+++ test/jdk/java/util/Calendar/bug4100311.java
@@ -23,31 +23,30 @@
 
 /*
  * @test
- * @bug 4028518
- * @summary Ensure cloned GregorianCalendar is unchanged when modifying its original.
- * @run junit ClonedGregorianUnchanged
+ * @bug 4100311
+ * @summary Ensure set(DAY_OF_YEAR, 1) works.
+ * @run junit bug4100311
  */
 
+import java.util.Calendar;
 import java.util.GregorianCalendar;
-
-import static java.util.Calendar.DAY_OF_MONTH;
+import java.util.Date;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class ClonedGregorianUnchanged {
+public class bug4100311 {
 
-    /*
-     * Ensure modifying the original GregorianCalendar does not
-     * modify the cloned one as well
-     */
+    // GregorianCalendar should be able to date to january 1st properly
+    @SuppressWarnings("deprecation")
     @Test
-    public void clonedShouldNotChangeOriginalTest() {
-        GregorianCalendar cal1 = new GregorianCalendar() ;
-        GregorianCalendar cal2 = (GregorianCalendar) cal1.clone() ;
-        cal1.add(DAY_OF_MONTH, 1) ;
-        assertNotEquals(cal1.get(DAY_OF_MONTH), cal2.get(DAY_OF_MONTH),
-                "Cloned calendar should not have same value as original");
+    public void dayOfYearIsOneTest() {
+        GregorianCalendar cal = new GregorianCalendar();
+        cal.set(Calendar.YEAR, 1997);
+        cal.set(Calendar.DAY_OF_YEAR, 1);
+        Date d = cal.getTime();
+        assertEquals(0, d.getMonth(), "Date: "+d+" isn't January 1st");
+        assertEquals(1, d.getDate(),"Date: "+d+" isn't January 1st");
     }
 }
