+++ test/jdk/java/util/Calendar/bug4243802.java
@@ -26,7 +26,7 @@
  * @bug 4243802
  * @summary confirm that Calendar.setTimeInMillis() and getTimeInMillis()
  *          can be called from a user program. They used to be protected methods.
- * @run junit SetTimeAndGetTimeInMillis
+ * @run junit bug4243802
  */
 
 import java.util.Calendar;
@@ -40,7 +40,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class SetTimeAndGetTimeInMillis {
+public class bug4243802 {
 
     private static final TimeZone savedTz = TimeZone.getDefault();
     private static final Locale savedLocale = Locale.getDefault();
