+++ test/jdk/java/util/Calendar/bug4372743.java
@@ -25,7 +25,7 @@
  * @test
  * @bug 4372743
  * @summary test that checks transitions of ERA and YEAR which are caused by add(MONTH).
- * @run junit AddMonthShouldTransitionEra
+ * @run junit bug4372743
  */
 
 import java.util.GregorianCalendar;
@@ -59,7 +59,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class AddMonthShouldTransitionEra {
+public class bug4372743 {
 
     private static final TimeZone savedTz = TimeZone.getDefault();
